backend/requirements.txt

python-dotenv
groq
httpx